Gui now displays the correct data Gui can now swap between feeds
